+++ scripts/helperFunctions.js
@@ -275,5 +275,8 @@ export let chris = {
 		let resistant = chris.checkTrait(actor, 'dr', type);
 		if (resistant) total = math.floor(total / 2);
 		return total;
+	},
+	'getEffectCastLevel': function _getEffectCastLevel(effect) {
+		return effect.flags['midi-qol']?.castData?.castLevel;
 	}
 };