+++ scripts/macros/monsterFeatures/dybbuk/dybbuk.js
@@ -1,0 +1,4 @@
+import {tentacle} from './tentacle.js';
+export let dybbuk = {
+    'tentacle': tentacle
+}