+++ scripts/macros/monsterFeatures/elderBrain/elderBrain.js
@@ -1,0 +1,6 @@
+import {breakConcentration} from './breakConcentration.js';
+import {psychicLink} from './psychicLink.js';
+export let elderBrain = {
+    'breakConcentration': breakConcentration,
+    'psychicLink': psychicLink
+}